+++ spring-cloud-starter-tencent-metadata-transfer/src/main/java/com/tencent/cloud/metadata/concurrent/executor/MetadataExecutors.java
@@ -104,7 +104,6 @@ public final class MetadataExecutors {
 	}
 
 	private MetadataExecutors() {
-
 	}
 
 }
