+++ spring-cloud-starter-tencent-metadata-transfer/src/test/java/com/tencent/cloud/metadata/concurrent/executor/MetadataExecutorServiceTest.java
@@ -124,4 +124,5 @@ public class MetadataExecutorServiceTest {
 	@SpringBootApplication
 	protected static class TestApplication {
 	}
+
 }
