+++ spring-cloud-starter-tencent-metadata-transfer/src/test/java/com/tencent/cloud/metadata/concurrent/executor/MetadataScheduledExecutorServiceTest.java
@@ -85,4 +85,5 @@ public class MetadataScheduledExecutorServiceTest {
 	@SpringBootApplication
 	protected static class TestApplication {
 	}
+
 }
